Added service to retrieve model as zipped JSON

+++ lib/m4db_serverside/rest/m4db_readonly_web/get_model_tecplot_file.py
@@ -138,12 +138,72 @@ class GetModelJSONFile:
             return
 
         # If the model exists but the json doesn't exist, then create the JSON
-        tecplot_data = read_tecplot(model_path_file, jsonify=True)
-        with open(json_path_file, "w") as fout:
-            fout.write(json.dumps(tecplot_data, sort_keys=True, indent=4))
+        if not os.path.exists(json_path_file):
+            tecplot_data = read_tecplot(model_path_file, jsonify=True)
+            with open(json_path_file, "w") as fout:
+                fout.write(json.dumps(tecplot_data, sort_keys=True, indent=4))
 
         # Return the stream
         if os.path.exists(json_path_file):
             resp.content_type = mimetypes.guess_type(json_path_file)[0]
             resp.stream = io.open(json_path_file, "rb")
             resp.content_length = os.path.getsize(json_path_file)
+
+
+class GetModelJSONFileZip:
+    r"""
+    Falcon service to retrieve a model JSON zipped file.
+    """
+
+    def on_get(self, req, resp, unique_id):
+        r"""
+        Retrieve the information associated with model object that has the given unique ID
+        as a zipped JSON file.
+        :param req: Falcon request object.
+        :param resp: Falcon response object.
+        :param unique_id: unique identifier of the model.
+        :return: None.
+        """
+        config = read_config_from_environ()
+
+        model_path_file = os.path.join(
+            config["file_root"],
+            global_vars.model_directory_name,
+            uid_to_dir(unique_id),
+            global_vars.magnetization_tecplot_file_name
+        )
+
+        json_path_file = os.path.join(
+            config["file_root"],
+            global_vars.model_directory_name,
+            uid_to_dir(unique_id),
+            global_vars.magnetization_json_file_name
+        )
+
+        zipped_json_path_file = os.path.join(
+            config["file_root"],
+            global_vars.model_directory_name,
+            uid_to_dir(unique_id),
+            global_vars.magnetization_json_zip_file_name
+        )
+
+        # If the model exists but the json doesn't exist, then create the JSON
+        if not os.path.exists(json_path_file):
+            tecplot_data = read_tecplot(model_path_file, jsonify=True)
+            with open(json_path_file, "w") as fout:
+                fout.write(json.dumps(tecplot_data, sort_keys=True, indent=4))
+
+        # The JSON zip files doesn't exist then create it
+        if os.path.exists(json_path_file):
+            if not os.path.exists(zipped_json_path_file):
+                # ... create the zip
+                zf = zipfile.ZipFile(zipped_json_path_file, "w", zipfile.ZIP_DEFLATED)
+                # Within the zip file, use the unique id
+                zf.write(json_path_file, "{}.json".format(unique_id))
+                zf.close()
+
+        # Return the stream
+        if os.path.exists(zipped_json_path_file):
+            resp.content_type = mimetypes.guess_type(zipped_json_path_file)[0]
+            resp.stream = io.open(zipped_json_path_file, "rb")
+            resp.content_length = os.path.getsize(zipped_json_path_file)
